backend/upload_function/requirements.txt
azure-functions
azure-ai-formrecognizer
azure-storage-blob
openai 